+++ DI-SourceGenerator/Attributes/GetByNameAttribute.cs
@@ -3,5 +3,8 @@
 namespace LurkingNinja.Attributes
 {
     [AttributeUsage(AttributeTargets.Field | AttributeTargets.Property)]
-    public class Get : Attribute {}
+    public class GetByName : Attribute
+    {
+        public GetByName(string gameObjectName) {}
+    }
 }