+++ DI-SourceGenerator/Attributes/GetByTagAttribute.cs
@@ -3,5 +3,8 @@
 namespace LurkingNinja.Attributes
 {
     [AttributeUsage(AttributeTargets.Field | AttributeTargets.Property)]
-    public class Get : Attribute {}
+    public class GetByTag : Attribute
+    {
+        public GetByTag(string tag) {}
+    }
 }