+++ DI-SourceGenerator/Attributes/InjectInPlay.cs
@@ -3,5 +3,5 @@
 namespace LurkingNinja.Attributes
 {
     [AttributeUsage(AttributeTargets.Field | AttributeTargets.Property)]
-    public class Get : Attribute {}
+    public class InjectInPlay : Attribute {}
 }