+++ Packages/com.lurking-ninja.dependency/Runtime/Placeholder.cs
@@ -1,0 +1,13 @@
+using UnityEngine;
+
+namespace LurkingNinja.Dependency
+{
+    public class Placeholder : MonoBehaviour
+    {
+        // Awake is called before the first frame update
+        private void Awake()
+        {
+        
+        }
+    }
+}