+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 All notable changes to this project will be documented in this file.
-## [0.0.1] - 2024-01-01
+## [0.0.1] - 2024-01-03
 ### Added
-### Changed
-### Removed
+- Get attribute
+- GetByName attribute
+- GetByTag attribute
+- GetInAssets attribute
+- GetInChildren attribute
+- GetInParent attribute
+- IgnoreSelf attribute
+- IncludeInactive attribute
+- InjectInPlay attribute
+- SkipNullCheck attribute
+### TODO
+- FindWithTag attribute